New release - from 0.1.0 to 0.2.0:  - added Dart null-safety  - added a Mac runner for the example app

+++ example/lib/_point_ext.dart
@@ -4,7 +4,7 @@ import 'dart:ui';
 
 extension FlutterPoint on Point {
   Offset toOffset() {
-    return Offset(this.x, this.y);
+    return Offset(this.x.toDouble(), this.y.toDouble());
   }
 }
